requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
openai>=1.12.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
flask-socketio==5.6.1
gunicorn==21.2.0
qrcode[pil]==7.4.2
Pillow>=9.5.0
eventlet==0.33.3